+++ scripts/rollup/shims/react-native/ReactNativeFeatureFlags.js
@@ -6,14 +6,14 @@
  * LICENSE file in the root directory of this source tree. An additional grant
  * of patent rights can be found in the PATENTS file in the same directory.
  *
- * @providesModule ReactErrorUtils
+ * @providesModule ReactNativeFeatureFlags
+ * @flow
  */
 
 'use strict';
 
-const {
-  __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,
-} = require('ReactNative');
+var ReactNativeFeatureFlags = {
+  useFiber: false,
+};
 
-module.exports =
-  __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.ReactErrorUtils;
+module.exports = ReactNativeFeatureFlags;
