+++ scripts/rollup/shims/react-native/takeSnapshot.js
@@ -6,8 +6,7 @@
  * LICENSE file in the root directory of this source tree. An additional grant
  * of patent rights can be found in the PATENTS file in the same directory.
  *
- * @providesModule NativeMethodsMixin
- * @flow
+ * @providesModule takeSnapshot
  */
 
 'use strict';
@@ -16,8 +15,5 @@ const {
   __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,
 } = require('ReactNative');
 
-import type {NativeMethodsMixinType} from 'ReactNativeTypes';
-
-const {NativeMethodsMixin} = __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED;
-
-module.exports = ((NativeMethodsMixin: any): $Exact<NativeMethodsMixinType>);
+module.exports =
+  __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.takeSnapshot;
