+++ src/renderers/native/ReactNativeFeatureFlags.js
@@ -12,8 +12,10 @@
 
 'use strict';
 
+// Read from process.env in order to support Rollup flat bundles.
+// Jest test script will also write this value for Fiber tests.
 var ReactNativeFeatureFlags = {
-  useFiber: false,
+  useFiber: process.env.REACT_NATIVE_USE_FIBER,
 };
 
 module.exports = ReactNativeFeatureFlags;
