+++ src/renderers/native/ReactNativeFiberHostComponent.js
@@ -23,9 +23,9 @@ import type {
   MeasureInWindowOnSuccessCallback,
   MeasureLayoutOnSuccessCallback,
   MeasureOnSuccessCallback,
-  NativeMethodsInterface,
-} from 'NativeMethodsMixinUtils';
-import type {Instance} from 'ReactNativeFiber';
+  NativeMethodsMixinType,
+} from 'ReactNativeTypes';
+import type {Instance} from 'ReactNativeFiberRenderer';
 import type {
   ReactNativeBaseComponentViewConfig,
 } from 'ReactNativeViewConfigRegistry';
@@ -37,7 +37,7 @@ import type {
  * ReactNativeFiber depends on this component and NativeMethodsMixin depends on
  * ReactNativeFiber).
  */
-class ReactNativeFiberHostComponent implements NativeMethodsInterface {
+class ReactNativeFiberHostComponent {
   _children: Array<Instance | number>;
   _nativeTag: number;
   viewConfig: ReactNativeBaseComponentViewConfig;
@@ -97,5 +97,8 @@ class ReactNativeFiberHostComponent implements NativeMethodsInterface {
     );
   }
 }
+
+// eslint-disable-next-line no-unused-expressions
+(ReactNativeFiberHostComponent.prototype: NativeMethodsMixinType);
 
 module.exports = ReactNativeFiberHostComponent;
